server/entity: Modernize helpers and tidy shared code
