+++ server/entity/arrow.go
@@ -45,14 +45,6 @@ var arrowConf = ProjectileBehaviourConfig{
 	SurviveBlockCollision: true,
 }
 
-// boolByte returns 1 if the bool passed is true, or 0 if it is false.
-func boolByte(b bool) uint8 {
-	if b {
-		return 1
-	}
-	return 0
-}
-
 // ArrowType is a world.EntityType implementation for Arrow.
 var ArrowType arrowType
