+++ server/entity/ent.go
@@ -280,3 +280,11 @@ func (e *Ent) checkPortalInsiders() bool {
 	}
 	return false
 }
+
+// boolByte returns 1 if the bool passed is true, or 0 if it is false.
+func boolByte(b bool) uint8 {
+	if b {
+		return 1
+	}
+	return 0
+}
